Removed TypeError catching and retaining Object#class
Object#class is no longer defined on Ronad::Monad. This was causing a
bug when combined with TypeError-catching delegating to method_missing.
TypeError-catching could have remained by implementing a
Ronad::Monad#from_value that would hardcode the actual class (since
Object#class was not available). This would then prevent a recursive
loop of trying to create a new monad via `self.class.new` (as self.class
would then also try to create a new Monad).

However, TypeError catching doesn't actually provide that much value.

    Maybe("6") + 2 #=> Maybe(nil)

This is misleading as the resulting Maybe(nil) was not caused by a nil
operation but rather TypeError. This nil would then proliferate
throughout the rest of the Maybe's lifecycle and tracing it is
completely non-obvious.