JustOne
Created a new monad that take multiple values and return the first
non-nil value.

Essentially the same as chaining with `||` however since `||` is a
method, chaining this with Maybe will short the express to always be
nil:

    (Maybe(nil) || Maybe(2)).value #=> nil

This can be circumvented by getting the value first:

    ~Maybe(nil) || ~Maybe(2)

However if you want to apply a common interface to a monad and just
interact with the one that is not nil a JustOne can be used:

    jo = JustOne(
      Maybe(params[:update_params])[:plan],
      Maybe(params[:params])[:plan],
      Maybe(params[:plan])
    )

This case documents that one of these three cases will have a value
(otherwise it will throw an error). Continuing from the previous
example, the monad can be used before evaluating the value:

    is_single = jo =~ /^single_/

    if ~is_single
      #...
    end